feat: add schema for note model
