--- src/api/notes/entities/note.schema.ts
@@ -1,21 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { HydratedDocument } from 'mongoose';
-
-export type NoteDocument = HydratedDocument<Note>;
-
-@Schema()
-export class Note {
-  @Prop({ required: true })
-  title: string;
-
-  @Prop({ required: true })
-  content: string;
-
-  @Prop({ required: true })
-  createdAt: string;
-
-  @Prop()
-  updatedAt: string;
-}
-
-export const NoteSchema = SchemaFactory.createForClass(Note);
